Extracted Un-/Locked padding compontents and specify PropTypes shape
